Validate that the user is an administrator using a token

+++ pages/api/admin/options/delete.ts
@@ -1,11 +1,14 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import deleteOptions from "./services/deleteOptions";
+import validateAdminUser from "../validateAdminUser";
 
 export default async function (req: NextApiRequest, res: NextApiResponse) {
     try {
         const { method } = req
         const { id } = req.body
+        const { authorization } = req.headers;
         if (method == 'DELETE') {
+            await validateAdminUser(authorization as string, 'eliminar la configuración')
             await deleteOptions(id)
             res.status(200).json({ message: "Configuración eliminada correctamente" })
         } else {
